Filter Duckling output to find specific parts of dates
This works by changing the reference times sent to Duckling to see if
the output of parsing the dates changes. This is because the design
decisions of Duckling are such that it favours upcoming dates over past
ones and it may output multiple possible dates if there are missing
components such as a missing year.